[man's voice] Once upon a time
in a kingdom far, far away,

the king and queen were blessed

with a beautiful baby girl.

And throughout the land,

everyone was happy...

until the sun went down

and they saw that their daughter was

cursed with a frightful enchantment

that took hold each and every night.

Desperate, they sought the help

of a fairy godmother

who had them lock the young princess

away in a tower,

there to await the kiss...

of the handsome Prince Charming.

[horse whinnies]

It was he who would chance

the perilous journey

through blistering cold

and scorching desert

traveling for many days and nights,

risking life and limb

to reach the Dragon's keep.

[crows caw]

For he was the bravest,

and most handsome...

in all the land.

And it was destiny that his kiss

would break the dreaded curse.

He alone would climb to the highest room

of the tallest tower